# Improve perf and memory usage with reuse cache by slicing inputs till token idx for 1st token generation.