--- a/examples/text-generation/run_generation.py
+++ b/examples/text-generation/run_generation.py
@@ -202,6 +202,8 @@ def main():
 
     if not args.use_hpu_graphs:
         args.limit_hpu_graphs = False
+        args.reuse_cache = False
+
     # Device is HPU
     args.device = "hpu"
     import habana_frameworks.torch.hpu as torch_hpu