Reuse cache pre allocates the KV cache. This can be further leveraged to slice the inputs till the token idx for the 1st token inference and improve perf and memory usage by running till the required seq len only rather than running full padded static input (input + max new tokens length). Memory usage and perf will vary according to the actual input seq len but will be better than running full static shapes.

Use torch.zeros to initialize KV cache rather than uninitialized torch.empty.

Also fix bool dtypes default values.

# What does this PR do?

<!--
Congratulations! You've made it this far! You're not quite done yet though.

Once merged, your PR is going to appear in the release notes with the title you set, so make sure it's a great title that fully reflects the extent of your awesome contribution.

Then, please replace this with a description of the change and which issue is fixed (if applicable). Please also include relevant motivation and context. List any dependencies (if any) that are required for this change.

Once you're done, someone will review your PR shortly (see the section "Who can review?" below to tag some potential reviewers). They may suggest changes to make the code even better. If no one reviewed your PR after a week has passed, don't hesitate to post a new comment @-mentioning the same persons---sometimes notifications get lost.
-->

<!-- Remove if not applicable -->

Fixes # (issue)


## Before submitting
- [ ] This PR fixes a typo or improves the docs (you can dismiss the other checks if that's the case).
- [ ] Did you make sure to update the documentation with your changes?
- [ ] Did you write any new necessary tests?