@puneeshkhanna 2023-09-25
Llama 7B perf results without and with fix as below:
python ../gaudi_spawn.py --use_deepspeed --world_size 1 run_generation.py --model_name_or_path /software/data/llama_inference/Llama-2-7b-hf/ --max_new_tokens 100 --bf16 --n_iterations 3 --use_hpu_graphs --use_kv_cache --batch_size 1 --limit_hpu_graphs --warmup 1 --attn_softmax_bf16 --trim_logits --reuse_cache
Without fix
----------------------------------------------------------------------
Throughput (including tokenization) = 131.44301810809804 tokens/second
Memory allocated                    = 12.62 GB
Max memory allocated                = 12.63 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 3.5840961140056606 seconds
----------------------------------------------------------------------
With fix
----------------------------------------------------------------------
Throughput (including tokenization) = 130.75967053114542 tokens/second
Memory allocated                    = 12.61 GB
Max memory allocated                = 12.62 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 2.1798126350040548 seconds
----------------------------------------------------------------------

python ../gaudi_spawn.py --use_deepspeed --world_size 1 run_generation.py --model_name_or_path /software/data/llama_inference/Llama-2-7b-hf/ --max_new_tokens 1000 --bf16 --n_iterations 3 --use_hpu_graphs --use_kv_cache --batch_size 1 --limit_hpu_graphs --warmup 1 --attn_softmax_bf16 --trim_logits --reuse_cache
Without fix
---------------------------------------------------------------------
Throughput (including tokenization) = 138.1225755548041 tokens/second
Memory allocated                    = 13.19 GB
Max memory allocated                = 13.23 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 10.586436815996422 seconds
---------------------------------------------------------------------
With fix
----------------------------------------------------------------------
Throughput (including tokenization) = 139.05396023775887 tokens/second
Memory allocated                    = 13.08 GB
Max memory allocated                = 13.1 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 8.58467547201144 seconds
----------------------------------------------------------------------

python ../gaudi_spawn.py --use_deepspeed --world_size 1 run_generation.py --model_name_or_path /software/data/llama_inference/Llama-2-7b-hf/ --max_new_tokens 100 --bf16 --n_iterations 3 --use_hpu_graphs --use_kv_cache --batch_size 128 --limit_hpu_graphs --warmup 1 --attn_softmax_bf16 --trim_logits --reuse_cache
Without fix
---------------------------------------------------------------------
Throughput (including tokenization) = 7084.259833894209 tokens/second
Memory allocated                    = 20.68 GB
Max memory allocated                = 22.21 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 17.2672253119963 seconds
---------------------------------------------------------------------
With fix
---------------------------------------------------------------------
Throughput (including tokenization) = 9801.250572694065 tokens/second
Memory allocated                    = 19.63 GB
Max memory allocated                = 20.96 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 3.1736303310026415 seconds
---------------------------------------------------------------------

python ../gaudi_spawn.py --use_deepspeed --world_size 1 run_generation.py --model_name_or_path /software/data/llama_inference/Llama-2-7b-hf/ --max_new_tokens 1000 --bf16 --n_iterations 3 --use_hpu_graphs --use_kv_cache --batch_size 32 --limit_hpu_graphs --warmup 1 --attn_softmax_bf16 --trim_logits --reuse_cache
Without fix
----------------------------------------------------------------------
Throughput (including tokenization) = 1865.7528637898442 tokens/second
Memory allocated                    = 31.03 GB
Max memory allocated                = 31.83 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 49.80568245299219 seconds
----------------------------------------------------------------------
With fix
----------------------------------------------------------------------
Throughput (including tokenization) = 2022.6046518376145 tokens/second
Memory allocated                    = 28.42 GB
Max memory allocated                = 29.18 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 17.86547275799967 seconds
----------------------------------------------------------------------