@puneeshkhanna 2023-09-25
Llama 13B perf results without and with fix as below:
python ../gaudi_spawn.py --use_deepspeed --world_size 1 run_generation.py --model_name_or_path /software/data/llama_inference/Llama-2-13b-hf/ --max_new_tokens 100 --bf16 --n_iterations 3 --use_hpu_graphs --use_kv_cache --batch_size 1 --limit_hpu_graphs --warmup 1 --attn_softmax_bf16 --trim_logits --reuse_cache
Without fix
Throughput (including tokenization) = 73.41020172978796 tokens/second
Memory allocated                    = 24.35 GB
Max memory allocated                = 24.36 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 4.884657262999099 seconds
With fix
Throughput (including tokenization) = 73.44032878995496 tokens/second
Memory allocated                    = 24.34 GB
Max memory allocated                = 24.35 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 3.37153663499339 seconds

python ../gaudi_spawn.py --use_deepspeed --world_size 1 run_generation.py --model_name_or_path /software/data/llama_inference/Llama-2-13b-hf/ --max_new_tokens 1000 --bf16 --n_iterations 3 --use_hpu_graphs --use_kv_cache --batch_size 1 --limit_hpu_graphs --warmup 1 --attn_softmax_bf16 --trim_logits --reuse_cache
Without fix
Throughput (including tokenization) = 75.7109614095682 tokens/second
Memory allocated                    = 25.2 GB
Max memory allocated                = 25.25 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 18.432767632009927 seconds
With fix
Throughput (including tokenization) = 76.26934607236079 tokens/second
Memory allocated                    = 25.06 GB
Max memory allocated                = 25.08 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 15.193954528993345 seconds

python ../gaudi_spawn.py --use_deepspeed --world_size 1 run_generation.py --model_name_or_path /software/data/llama_inference/Llama-2-13b-hf/ --max_new_tokens 100 --bf16 --n_iterations 3 --use_hpu_graphs --use_kv_cache --batch_size 64 --limit_hpu_graphs --warmup 1 --attn_softmax_bf16 --trim_logits --reuse_cache
Without fix
Throughput (including tokenization) = 2865.2254885524662 tokens/second
Memory allocated                    = 30.45 GB
Max memory allocated                = 31.22 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 17.60540542700619 seconds
With fix
Throughput (including tokenization) = 3633.9957034341414 tokens/second
Memory allocated                    = 29.76 GB
Max memory allocated                = 30.42 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 4.64509762400121 seconds

python ../gaudi_spawn.py --use_deepspeed --world_size 1 run_generation.py --model_name_or_path /software/data/llama_inference/Llama-2-13b-hf/ --max_new_tokens 1000 --bf16 --n_iterations 3 --use_hpu_graphs --use_kv_cache --batch_size 16 --limit_hpu_graphs --warmup 1 --attn_softmax_bf16 --trim_logits --reuse_cache
Without fix
Throughput (including tokenization) = 763.3575211375546 tokens/second
Memory allocated                    = 38.5 GB
Max memory allocated                = 38.92 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 49.66154878999805 seconds
With fix
Throughput (including tokenization) = 812.995454113312 tokens/second
Memory allocated                    = 36.65 GB
Max memory allocated                = 37.03 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 22.37336895200133 seconds