@puneeshkhanna 2023-09-25
Llama70B results:

python ../gaudi_spawn.py --use_deepspeed --world_size 8 run_generation.py --model_name_or_path /software/data/llama_inference/Llama-2-70b-hf/ --max_new_tokens 100 --bf16 --n_iterations 3 --use_hpu_graphs --use_kv_cache --batch_size 2 --limit_hpu_graphs --warmup 1 --attn_softmax_bf16 --trim_logits --reuse_cache
Without fix
Throughput (including tokenization) = 110.62860953193044 tokens/second
Memory allocated                    = 16.96 GB
Max memory allocated                = 17.37 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 3.4168114269996295 seconds
With fix
Throughput (including tokenization) = 109.80677493423512 tokens/second
Memory allocated                    = 16.95 GB
Max memory allocated                = 16.98 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 3.253933574000257 seconds

python ../gaudi_spawn.py --use_deepspeed --world_size 8 run_generation.py --model_name_or_path /software/data/llama_inference/Llama-2-70b-hf/ --max_new_tokens 4000 --bf16 --n_iterations 3 --use_hpu_graphs --use_kv_cache --batch_size 2 --limit_hpu_graphs --warmup 1 --attn_softmax_bf16 --trim_logits --reuse_cache
Without fix
Throughput (including tokenization) = 116.59096866609697 tokens/second
Memory allocated                    = 18.87 GB
Max memory allocated                = 39.02 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 70.28810204200272 seconds
With fix
Throughput (including tokenization) = 117.39353441901102 tokens/second
Memory allocated                    = 17.54 GB
Max memory allocated                = 17.58 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 69.67671067500487 seconds

python ../gaudi_spawn.py --use_deepspeed --world_size 8 run_generation.py --model_name_or_path /software/data/llama_inference/Llama-2-70b-hf/ --max_new_tokens 100 --bf16 --n_iterations 3 --use_hpu_graphs --use_kv_cache --batch_size 40 --limit_hpu_graphs --warmup 1 --attn_softmax_bf16 --trim_logits --reuse_cache
Without fix
Throughput (including tokenization) = 1744.4332593278932 tokens/second
Memory allocated                    = 17.71 GB
Max memory allocated                = 31.47 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 3.674584093008889 seconds
With fix
Throughput (including tokenization) = 1873.2515938931053 tokens/second
Memory allocated                    = 17.33 GB
Max memory allocated                = 17.95 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 3.530354335991433 seconds

python ../gaudi_spawn.py --use_deepspeed --world_size 8 run_generation.py --model_name_or_path /software/data/llama_inference/Llama-2-70b-hf/ --max_new_tokens 4000 --bf16 --n_iterations 3 --use_hpu_graphs --use_kv_cache --batch_size 40 --limit_hpu_graphs --warmup 1 --attn_softmax_bf16 --trim_logits --reuse_cache
Without fix
Throughput (including tokenization) = 1241.8834557707662 tokens/second
Memory allocated                    = 50.08 GB
Max memory allocated                = 88.1 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 137.97351068799617 seconds
With fix
Throughput (including tokenization) = 1409.4239974630066 tokens/second
Memory allocated                    = 23.58 GB
Max memory allocated                = 24.4 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 115.37053568501142 seconds

python ../gaudi_spawn.py --use_deepspeed --world_size 8 run_generation.py --model_name_or_path /software/data/llama_inference/Llama-2-70b-hf/ --max_new_tokens 1000 --bf16 --n_iterations 3 --use_hpu_graphs --use_kv_cache --batch_size 16 --limit_hpu_graphs --warmup 1 --attn_softmax_bf16 --trim_logits --reuse_cache --prompt "<300 tokens>"
Without fix
Throughput (including tokenization) = 778.5468898740977 tokens/second
Memory allocated                    = 19.79 GB
Max memory allocated                = 41.1 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 22.690399918996263 seconds
With fix
Throughput (including tokenization) = 831.5857981807274 tokens/second
Memory allocated                    = 18.26 GB
Max memory allocated                = 34.62 GB
Total memory available              = 94.46 GB
Graph compilation duration          = 21.228477083001053 seconds